+++ frontend/src/entry/model/registry/VulnerabilityResult.tsx
@@ -15,34 +15,8 @@ import { grey } from '@mui/material/colors'
 import { useTheme } from '@mui/material/styles'
 import { useCallback, useEffect, useEffectEvent, useMemo, useState } from 'react'
 import Link from 'src/Link'
-import { ArtefactScanState, ImageTagResult, ScanState, SeverityLevel } from 'types/types'
+import { ArtefactScanState, ImageTagResult, ScanState } from 'types/types'
 import { plural } from 'utils/stringUtils'
-
-const severities = [
-  SeverityLevel.CRITICAL,
-  SeverityLevel.HIGH,
-  SeverityLevel.MEDIUM,
-  SeverityLevel.LOW,
-  SeverityLevel.UNKNOWN,
-]
-
-function highestVulnerabilityPlatform(scanResults: ImageTagResult[]): string | undefined {
-  //Can we add the severities loop here? and go through each result.
-  const sortedResults = scanResults.sort((a, b) => {
-    let res = 0
-    for (let i = 0; i < severities.length; i++) {
-      if (a.severityCounts[severities[i]] < b.severityCounts[severities[i]]) {
-        res = -1
-        break
-      } else if (a.severityCounts[severities[i]] > b.severityCounts[severities[i]]) {
-        res = 1
-        break
-      }
-    }
-    return res
-  })
-  return sortedResults[sortedResults.length - 1].platform
-}
 
 function selectedCurrentResult(
   multiarchitecture: boolean,
@@ -58,7 +32,7 @@ function selectedCurrentResult(
 
 type VulnerabilityResultProps = {
   scanResults: ImageTagResult[] | undefined
-  platforms?: (string | undefined)[] | undefined
+  platforms?: (string | undefined)[]
   warningOnly?: boolean
   detailedViewUrlPrefix?: string
 }
@@ -77,7 +51,7 @@ export default function VulnerabilityResult({
 
   const multiarchitecture = (!!platforms && !!scanResults) ?? false
   const [currentPlatform, setCurrentPlatform] = useState<string | undefined>(
-    multiarchitecture && scanResults ? highestVulnerabilityPlatform(scanResults) : undefined,
+    multiarchitecture && platforms ? platforms[0] : undefined,
   )
   const [currentResult, setCurrentResult] = useState<ImageTagResult | undefined>(
     scanResults ? selectedCurrentResult(multiarchitecture, currentPlatform, scanResults) : undefined,
